refactor(git): rename git-utils to git-provider; migrate tests and update imports
Introduce git-provider module replacing git-utils.\n\nChanges:\n- Rename and relocate: git-utils.mjs -> git-provider/git-provider.mjs (preserves all existing exports)\n- Add colocated test suite at src/git-provider/tests/git-provider.test.mjs\n- Update all imports (gen-pr, gen-mr, merge-request-generator, prompt-generator, validation) to new path\n- Adjust jest testMatch ordering to include new provider test directory first\n- Migrate and adapt tests (prompt-generator, integration) and mock paths\n- Update documentation (README, TESTING.md, PULL_REQUEST_UPDATE_FEATURE.md) to reflect new naming\n- Keep legacy placeholder test file (src/tests/git-utils.test.mjs) marked skipped to avoid breakage (scheduled for removal)\n\nRationale:\nAlign naming with provider pattern and enable clearer module ownership. No functional logic changes intended.
